requirements.txt
requests~=2.25.1
joblib~=1.1.1
pandas~=1.5.2
scikit-learn~=1.2.0
xgboost~=1.7.3
uWSGI==2.0.20
flask==2.1.3
python-dotenv==0.20.0
